Initialize a Jackcess Database object instance for interacting with MS Access database files

Initialize a Jackcess Database object instance for interacting with MS Access database files

Usage

Database(
  dsn,
  read_only = FALSE,
  file_format = c("V2019", "V2016", "V2010", "V2007", "V2003", "V2000", "V1997",
    "MSISAM", "GENERIC_JET4"),
  overwrite = FALSE
)

Arguments

dsn

character. MS Access database source name (file path).

read_only

logical Open connection as read-only? Default FALSE

file_format

character. One of the following: "GENERIC_JET4", "MSISAM", "V1997", "V2000", "V2003", "V2007", "V2010", "V2016", "V2019". See details.

overwrite

logical. overwrite existing file? Default FALSE

Value

A com.healthmarketscience.jackcess.Database object

Details

"GENERIC_JET4" - A database which was most likely created programmatically

"MSISAM" - A database which was created by MS Money

"V1997" - A database which was created by MS Access 97

"V2000" - A database which was created by MS Access 2000

"V2003" - A database which was created by MS Access 2002/2003

"V2007" - A database which was created by MS Access 2007

"V2010" - A database which was created by MS Access 2010+

"V2016" - A database which was created by MS Access 2016+

"V2019" - A database which was created by MS Access 2019+ (Office 365)